% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean data from NOAA's Significant Earthquake Database}
\usage{
eq_clean_data(db)
}
\arguments{
\item{db}{A data.frame of the earthquake database with a columns "YEAR", "MONTH", "DAY", "LATITUE", "LONGITUDE", "LOCATION_NAME".}
}
\value{
A cleaned data.frame containing earthquake information
}
\description{
This function takes in data from NOAA's Significant Earthquake Database and cleans it into a form that the other functions in the AC.R.Capstone package take as input.
\itemize{
  \item LOCATION_NAME has country removed and is put in title case
  \item YEAR, MONTH, DAY is combined into a single field DATE
  \item LONGITUDE and LATITUDE are converted to numeric variables
}
}
\examples{
\dontrun{
db <- AC.R.Capstone::NOAA_eq_data
tail(db)
clean_data <- eq_clean_data(db)
tail(clean_data)
}

}
