% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_plots.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML Labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data.frame that contains the columns LOCATION_NAME, EQ_PRIMARY, and TOTAL_DEATHS}
}
\value{
A Character string written in HTML code that can be used as a leaflet label
}
\description{
Used to create leaflet labels written in HTML that contain earthquake information. It will return a label for each row of the input data.frame.
}
\examples{
\dontrun{
  db <- AC.R.Capstone::NOAA_eq_data
  clean_data <- AC.R.Capstone::eq_clean_data(db)
  clean_data$label <- AC.R.Capstone::eq_create_label(clean_data)
}

}
