% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Name}
\usage{
eq_location_clean(db)
}
\arguments{
\item{db}{A data.frame of the eartquake database with a column called "LOCATION_NAME"}
}
\value{
An identical data.frame with the exception that the "LOCATION_NAME" is cleaned.
}
\description{
This function takes in a data.frame and returns it back with the column called "LOCATION_NAME" cleaned.
Cleaned means the country name has been removed and the location name has been put into title case.
}
\examples{
\dontrun{
db <- AC.R.Capstone::NOAA_eq_data
tail(db)
clean_data <- eq_location_clean(db)
tail(clean_data)
}

}
