% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_plots.R
\name{eq_map}
\alias{eq_map}
\title{Plot an interactive map of Earthquake Locations}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data.frame containing earthquake information including EQ_PRIMARY, LONGITUDE, and LATITUDE}

\item{annot_col}{A String that names the field in the earthquake data.frame to be used for annotating the earthquake points}
}
\value{
A leaflet htmlwigit that is is a interactive map which can be viewed in a browser.
}
\description{
Plots a leaflet map of the data with points at earthquake locations. Points can be clicked on to see additional information.
}
\examples{
\dontrun{
db <- AC.R.Capstone::NOAA_eq_data
clean_data <- eq_clean_data(db)
select_data <- clean_data \%>\%
 dplyr::filter(COUNTRY \%in\% c("ARGENTINA")) \%>\%
 dplyr::mutate(label = eq_create_label(.))
AC.R.Capstone::eq_map(select_data, "label")
}

}
