% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plots.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Using ggplot2::aes() to assign columns from data for use in the plot. Options relevant to this geom are:
\itemize{
  \item x - dates of points on the timeline
  \item xmin - date to start the timeline
  \item xmax - date to end the timeline
  \item y - variable to differentiate different lines of dates
  \item fill - variable to color the points by
  \item size - variable to determine the size of points
  \item alpha - the transparency of the points
}}

\item{data}{a data.frame that will be used for the plot}

\item{...}{Other parameters that are used by ggplot2::layer()}
}
\value{
A layer that can be added to a ggplot2 graph
}
\description{
This timeline plots lines for each y category (typically would put country here), with points indicating the dates of earthquakes.
The size of the point can be used for the magnitude of the earthquake and the color for the number of deaths.
To match the plot style of the assignment, add the theme AC.R.Capstone::theme_timeline to the plot.
}
\examples{
\dontrun{
db <- AC.R.Capstone::NOAA_eq_data
clean_data <- eq_clean_data(db)
select_data <- clean_data \%>\%
  dplyr::filter(COUNTRY \%in\% c("ARGENTINA", "BRAZIL", "COLOMBIA")) \%>\%
  dplyr::filter(DATE >= lubridate::ymd("1950-01-01")) \%>\%
  dplyr::filter(DATE <= lubridate::ymd("1999-12-31"))
ggplot2::ggplot(data = select_data) +
  AC.R.Capstone::geom_timeline(mapping = ggplot2::aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, fill = DEATHS),
                             xmin = lubridate::ymd("1950-01-01"), xmax = lubridate::ymd("1999-12-31")) +
  AC.R.Capstone::theme_timeline
}

}
