% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plots.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Add labels to Timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Using ggplot2::aes() to assign columns from data for use in the plot. Options relevant to this geom are:
\itemize{
  \item x - dates of points on the timeline
  \item label - the text field to use as a label
  \item y - variable to differentiate different lines of dates
  \item size - variable to determine the size of points
  \item n_max - the maximum number of points to label. Will label biggest points if the size variable is provided.
}}

\item{data}{a data.frame that will be used for the plot}

\item{...}{Other parameters that are used by ggplot2::layer()}
}
\value{
A layer that can be added to a ggplot2 graph
}
\description{
This geom can be used to add labels to points of a ggplot2 graph with a geom_timeline.
}
\examples{
\dontrun{
db <- AC.R.Capstone::NOAA_eq_data
clean_data <- eq_clean_data(db)
select_data <- clean_data \%>\%
  dplyr::filter(COUNTRY \%in\% c("ARGENTINA")) \%>\%
  dplyr::filter(DATE >= lubridate::ymd("1950-01-01")) \%>\%
  dplyr::filter(DATE <= lubridate::ymd("1999-12-31"))
ggplot2::ggplot(data = select_data) +
  AC.R.Capstone::geom_timeline(mapping = ggplot2::aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, fill = DEATHS),
                             xmin = lubridate::ymd("1950-01-01"), xmax = lubridate::ymd("1999-12-31")) +
  AC.R.Capstone::geom_timeline_label(mapping = ggplot2::aes(x = DATE, label = LOCATION_NAME)) +
  AC.R.Capstone::theme_timeline
}

}
