% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAA_eq_data}
\alias{NOAA_eq_data}
\title{Earthquake Data from NOAA}
\format{data.frame}
\source{
National Geophysical Data Center / World Data Service (NGDC/WDS): Significant Earthquake Database. National Geophysical Data Center, NOAA. doi:10.7289/V5TD9V7K
}
\usage{
NOAA_eq_data
}
\description{
Data from the NOAA downloaded on November 3, 2019
}
\examples{
\dontrun{
  NOAA_eq_data
}
}
\keyword{datasets}
