% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_plots.R
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for timeline plots}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 67.}
\usage{
theme_timeline
}
\description{
In order to have the timeline plots match the style of examples provided in the Mastering Software Development in R Course add this theme to the plot
}
\keyword{datasets}
